requirements.txt
click
csscompressor
htmlmin
Jinja2
Markdown
MarkupSafe
#pathtools
ply
PyYAML
#slimit
watchdog
webptools
